- Refactor to potentially support alternative caster object in prevision of creating "Tower".
Note: It seems like utilizing interface is a most. However, only using interface requires to reimplement a bunch of feature for each implementation. Ideally, we should "component" to implement the feature however it is hard due to how specific those feature gets. To remediate this situation you should normally use the Inverse Dependency Principle to invert the dependence. However doing so result in a strange pattern such as ICaster - ICasterOwner, IAttackable - IAttackableOwner, ITargateable - ITargateableOwner which is far from ideal.