- Refactor statistic workflow to use non-reflection. - Make the possibility of having "Statistic" span on multiple component from an AgentObject. - Make it possible to reference statistic directly from code without using StatisticReference. - Add a "Caster" component which handle the logic of casting a spell.
Note: Will attempt to redo the separation of concern with Statistic in mind. This way, we would be able to define a generic attackable. The issue with the previous implementation is that the number of interface was growing really fast, they were a lot of overlap between them, and some parameters does not make sense in some context. By example, from a single interface we will be able to access properties such as Health, MaxHealth and Defense without having to define an interface for those. It is also worth mentioning that we will also be able to reference properties such as IncreaseDamage that might have been considerably specific.