- Refactor Blocker + Animated in a "Component".
Note: The reason why the previous attempt to divide the object in diverse part failed was because of three reasons:
- We did not have a pivot class generic enough which make class such as Target cumbersome to use given that we had to find a way to expose the Faction of the Entity. (Now, we can simply check if the Entity is a AgentObject)
- We were not able to access specific attribute.(i.e. Current Health, Defense)
- We were not able to find an acceptable way to deal with the effect of a specific component such as Attackable. (Now, we listen to the result of an attack and then applied the effect such as losing health)

That being said, we do now even more realize the dynamic between generic implementation and specific implementation. While generic implementation adds a considerable amount of flexibility, it lacks the usability of a more concrete implementation. The best is to have a mixture of both were the specific can be express as generic when needed. By example, when considering the Statistic of an object, we can have specific statistic defined in code such as IsEngaged or CurrentHealth that are exposed in a more generic manner to the external element of the class. We choose to attempt to code generation, but reflection or/and function such as what was done previously work as well.