- Revert to original approach of using OrientedObject.
While it might be easier to create statistic from code with the older approach, it makes the "dynamic" part hard or impossible. If we want to define statistic as serialized data, we still need to recreate the whole original architecture. That being said, code generation could potentially be used to create the function but it seem a bit too much for the advantage we would get from it.