- Make Modifier a MonoBehaviour
Note: Using Unity MonoBehaviour prevent us from needing to create complicated factory when dealing with generic structure as we can copy the whole object. While this throw us back pretty far in development, we do not really care as it is a learning project where the end result is not necessary the final product. The older approach was not necessary bad, it simply required a considerable amount of code to be written in double given that we had to define every modifier behaviour from code. With the is new approach, we doing something similar to what we usually do with Projectile / Ability.